Broke out util code
