+++ server/src/main.rs
@@ -1,9 +1,6 @@
 mod auth;
+mod util;
 
-use argon2::{
-    password_hash::{PasswordHasher, Salt},
-    Argon2,
-};
 use auth::{login, regen_token, register_streamer};
 use axum::{
     extract,
@@ -24,6 +21,7 @@ use serde::{Deserialize, Serialize};
 use sqlx::postgres::{PgPool, PgPoolOptions};
 use std::env;
 use tower_http::cors::{Any, CorsLayer};
+use util::verify_hash;
 
 #[derive(Deserialize, Debug)]
 struct Viewer {
@@ -184,20 +182,6 @@ async fn push(
         body: Some("Success"),
         error: None,
     })
-}
-
-fn verify_hash(hashed_value: &str, unhashed_value: &str) -> bool {
-    let salt = hashed_value.split("$").collect::<Vec<&str>>()[4];
-
-    let pass = Argon2::default()
-        .hash_password(unhashed_value.as_bytes(), Salt::from_b64(salt).unwrap())
-        .unwrap();
-
-    if pass.to_string() == hashed_value {
-        return true;
-    }
-
-    return false;
 }
 
 async fn toggle_stream(
